moved service_fee and stake_per_node to settings
