+++ v0_4/src/user_stake.rs
@@ -1,12 +1,11 @@
 
 use crate::events::*;
-use crate::node_config::*;
 use crate::pause::*;
 use crate::settings::*;
 use crate::user_data::*;
 use crate::fund_transf_module::*;
 use crate::fund_view_module::*;
-use crate::node_activation::*;
+// use crate::settings::*;
 
 imports!();
